8.2.2

+++ com/baidu/tieba/service/FatalErrorService.java
@@ -12,9 +12,9 @@ import com.baidu.sapi2.SapiAccountManager;
 import com.baidu.tbadk.TbConfig;
 import com.baidu.tbadk.core.TbadkCoreApplication;
 import com.baidu.tbadk.core.util.TiebaStatic;
-import com.baidu.tbadk.core.util.av;
-import com.baidu.tbadk.core.util.m;
-import com.baidu.tbadk.core.util.z;
+import com.baidu.tbadk.core.util.at;
+import com.baidu.tbadk.core.util.l;
+import com.baidu.tbadk.core.util.y;
 import java.io.BufferedReader;
 import java.io.ByteArrayOutputStream;
 import java.io.File;
@@ -57,8 +57,8 @@ public class FatalErrorService extends BdBaseService {
 
     /* loaded from: classes.dex */
     private class a extends BdAsyncTask<String, Integer, String> {
-        private z aiN = null;
         Intent intent;
+        private y mNetwork = null;
 
         public a(Intent intent) {
             this.intent = intent;
@@ -135,12 +135,12 @@ public class FatalErrorService extends BdBaseService {
                                         }
                                         return;
                                     }
-                                    this.aiN = new z(String.valueOf(TbConfig.SERVER_ADDRESS) + str);
-                                    this.aiN.d("logfile", byteArray);
+                                    this.mNetwork = new y(String.valueOf(TbConfig.SERVER_ADDRESS) + str);
+                                    this.mNetwork.d("logfile", byteArray);
                                     if (!TextUtils.isEmpty(str2) && !"0".equals(str2)) {
-                                        this.aiN.n(FatalErrorService.ERROR_TYPE_KEY, str2);
+                                        this.mNetwork.n(FatalErrorService.ERROR_TYPE_KEY, str2);
                                     }
-                                    this.aiN.un();
+                                    this.mNetwork.ug();
                                     if (byteArrayOutputStream2 != null) {
                                         byteArrayOutputStream2.close();
                                         byteArrayOutputStream3 = null;
@@ -209,7 +209,7 @@ public class FatalErrorService extends BdBaseService {
                                         fileInputStream2 = fileInputStream;
                                     }
                                     try {
-                                        if (this.aiN.uI().vC().oH()) {
+                                        if (this.mNetwork.uC().vw().isRequestSuccess()) {
                                             if (z2) {
                                                 z(file);
                                             }
@@ -380,16 +380,16 @@ public class FatalErrorService extends BdBaseService {
             File[] listFiles;
             boolean z = true;
             try {
-                a(m.cZ(TbConfig.FATAL_ERROR_FILE), TbConfig.ERROR_UPLOAD_SERVER, "0", true, true);
-                a(m.cZ(TbConfig.LOG_ERROR_FILE), "c/s/clientlog", "0", false, false);
-                aYX();
-                if (!TbConfig.getVersion().equals(com.baidu.tbadk.core.sharedPref.b.tW().getString("native_crash_dump_version", ""))) {
-                    com.baidu.tbadk.core.sharedPref.b.tW().putString("native_crash_dump_version", TbConfig.getVersion());
+                a(l.cX(TbConfig.FATAL_ERROR_FILE), TbConfig.ERROR_UPLOAD_SERVER, "0", true, true);
+                a(l.cX(TbConfig.LOG_ERROR_FILE), "c/s/clientlog", "0", false, false);
+                baM();
+                if (!TbConfig.getVersion().equals(com.baidu.tbadk.core.sharedPref.b.tQ().getString("native_crash_dump_version", ""))) {
+                    com.baidu.tbadk.core.sharedPref.b.tQ().putString("native_crash_dump_version", TbConfig.getVersion());
                     z = false;
                 }
-                File cZ = m.cZ(TbConfig.FATAL_ERROR_NATIVE_DIR);
-                if (cZ != null) {
-                    for (File file : cZ.listFiles()) {
+                File cX = l.cX(TbConfig.FATAL_ERROR_NATIVE_DIR);
+                if (cX != null) {
+                    for (File file : cX.listFiles()) {
                         if (file.length() >= IjkMediaMeta.AV_CH_SIDE_RIGHT && z) {
                             A(file);
                             a(file, TbConfig.ERROR_UPLOAD_SERVER, FatalErrorService.ERROR_TYPE_NATIVE_C, true, true);
@@ -406,11 +406,11 @@ public class FatalErrorService extends BdBaseService {
             }
         }
 
-        private void aYX() {
-            File cZ = m.cZ(TbConfig.FATAL_ERROR_ALERT_FILE);
-            if (cZ != null) {
+        private void baM() {
+            File cX = l.cX(TbConfig.FATAL_ERROR_ALERT_FILE);
+            if (cX != null) {
                 try {
-                    BufferedReader bufferedReader = new BufferedReader(new InputStreamReader(new FileInputStream(cZ)));
+                    BufferedReader bufferedReader = new BufferedReader(new InputStreamReader(new FileInputStream(cX)));
                     StringBuffer stringBuffer = new StringBuffer();
                     while (true) {
                         String readLine = bufferedReader.readLine();
@@ -419,8 +419,8 @@ public class FatalErrorService extends BdBaseService {
                         } else {
                             String stringBuffer2 = stringBuffer.toString();
                             BdLog.i("sendLogForAlert log = " + stringBuffer2);
-                            com.baidu.adp.lib.stats.a.eI().p("alert_crash", stringBuffer2);
-                            cZ.delete();
+                            com.baidu.adp.lib.stats.a.eG().p("alert_crash", stringBuffer2);
+                            cX.delete();
                             return;
                         }
                     }
@@ -434,43 +434,43 @@ public class FatalErrorService extends BdBaseService {
 
         private void A(File file) {
             FileWriter fileWriter;
-            if (file != null && file.exists() && file.isFile()) {
+            if (file != null && file.exists() && file.isFile() && this.intent != null) {
                 try {
-                    if (this.intent != null) {
+                    fileWriter = new FileWriter(file, true);
+                    try {
                         try {
-                            fileWriter = new FileWriter(file, true);
-                            try {
-                                fileWriter.append("\n##TIEBA_NATIVE##\n");
-                                a(fileWriter, av.vj(), null);
-                                a(fileWriter, "version", TbConfig.getVersion());
-                                a(fileWriter, "model", Build.MODEL);
-                                a(fileWriter, "android_version", Build.VERSION.RELEASE);
-                                a(fileWriter, "android_sdk", String.valueOf(Build.VERSION.SDK_INT));
-                                a(fileWriter, "from", TbConfig.getFrom());
-                                a(fileWriter, "current_from", TbConfig.getCurrentFrom());
-                                a(fileWriter, SapiAccountManager.SESSION_UID, this.intent.getStringExtra(SapiAccountManager.SESSION_UID));
-                                a(fileWriter, "client_id", TbadkCoreApplication.getClientId());
-                                a(fileWriter, "imei", TbadkCoreApplication.m9getInst().getImei());
-                                a(fileWriter, "uname", this.intent.getStringExtra("uname"));
-                                fileWriter.append("\n##TIEBA_NATIVE_END##\n");
-                                o.a(fileWriter);
-                            } catch (Exception e) {
-                                e = e;
-                                e.printStackTrace();
-                                o.a(fileWriter);
-                            }
-                        } catch (Exception e2) {
-                            e = e2;
-                            fileWriter = null;
-                        } catch (Throwable th) {
-                            th = th;
-                            fileWriter = null;
+                            fileWriter.append("\n##TIEBA_NATIVE##\n");
+                            a(fileWriter, at.vd(), null);
+                            a(fileWriter, "version", TbConfig.getVersion());
+                            a(fileWriter, "model", Build.MODEL);
+                            a(fileWriter, "android_version", Build.VERSION.RELEASE);
+                            a(fileWriter, "android_sdk", String.valueOf(Build.VERSION.SDK_INT));
+                            a(fileWriter, "from", TbConfig.getFrom());
+                            a(fileWriter, "current_from", TbConfig.getCurrentFrom());
+                            a(fileWriter, SapiAccountManager.SESSION_UID, this.intent.getStringExtra(SapiAccountManager.SESSION_UID));
+                            a(fileWriter, "client_id", TbadkCoreApplication.getClientId());
+                            a(fileWriter, "imei", TbadkCoreApplication.m9getInst().getImei());
+                            a(fileWriter, "uname", this.intent.getStringExtra("uname"));
+                            fileWriter.append("\n##TIEBA_NATIVE_END##\n");
                             o.a(fileWriter);
-                            throw th;
+                        } catch (Exception e) {
+                            e = e;
+                            e.printStackTrace();
+                            o.a(fileWriter);
                         }
+                    } catch (Throwable th) {
+                        th = th;
+                        o.a(fileWriter);
+                        throw th;
                     }
+                } catch (Exception e2) {
+                    e = e2;
+                    fileWriter = null;
                 } catch (Throwable th2) {
                     th = th2;
+                    fileWriter = null;
+                    o.a(fileWriter);
+                    throw th;
                 }
             }
         }
@@ -490,8 +490,8 @@ public class FatalErrorService extends BdBaseService {
 
         @Override // com.baidu.adp.lib.asyncTask.BdAsyncTask
         public void cancel() {
-            if (this.aiN != null) {
-                this.aiN.eg();
+            if (this.mNetwork != null) {
+                this.mNetwork.ee();
             }
             FatalErrorService.this.mTask = null;
             super.cancel(true);
@@ -501,9 +501,8 @@ public class FatalErrorService extends BdBaseService {
         /* JADX DEBUG: Method merged with bridge method */
         /* JADX INFO: Access modifiers changed from: protected */
         @Override // com.baidu.adp.lib.asyncTask.BdAsyncTask
-        /* renamed from: ew */
         public void onPostExecute(String str) {
-            super.onPostExecute(str);
+            super.onPostExecute((a) str);
             FatalErrorService.this.mTask = null;
             FatalErrorService.this.stopSelf();
         }
